[#3] Move NotificationOpened event to proper namespace
